+++ src/index.html
@@ -20,28 +20,22 @@
             <div class="score-seciont">
                 <button type="refresh" id="refresh">Refresh</button>
                 <hr>
-                <table id="table">
-                    <tr>
-                        <th>Name</th>
-                        <th>Score</th>
-                    </tr>
-                    <tr>
-                        <td>data</td>
-                        <td>data</td>
-                    </tr>
-                </table>
+                <ul id="table">
+
+                </ul>
             </div>
             <div class="entry-section">
                 <form class="form">
                     <div class="name">
-                        <label class="label" id="name">Enter Name</label>
-                        <input type="text" class="input">
+                        <label class="label" >Enter Name</label>
+                        <input type="text" class="input" id="name">
                     </div>
                     <div class="score">
-                        <label class="label" id="score">Enter Score</label>
-                        <input type="text" class="input">
+                        <label class="label" >Enter Score</label>
+                        <input type="text" class="input" id="score">
                     </div>
                     <button type="submit" id="submit">Add Score</button>
+                    <p id="message"></p>
                 </form>
             </div>
         </div>
